backend/requirements.txt
fastapi
pydantic
pymongo
python-dotenv
passlib[bcrypt]
python-jose
